Add code to extract permanent files from UMinn DUMPPF records.
The University of Minnesota DUMPPF utility dumps a permanent file to
tape in a fairly straightforward format: 7700 table, 7400 table
containing the PF catalog entry, file contents in a continuous stream
of 60-bit words as returned via READCW.

The ARCHIVE utility uses a similar format, except the 7400 table is
empty.

Add support to extract DUMPPF records as I-format tape images in SIMH
format, as is already done with PFDUMP records.